requirements.txt
pandas~=2.1.0
pytest~=7.4.2
requests~=2.31.0
boto3~=1.28.49
botocore~=1.31.49
fake-useragent
tqdm~=4.66.1
asyncio~=3.4.3
newspaper3k
beautifulsoup4~=4.12.2
tldextract~=3.5.0